Provide a workaround for misconfigured Google Apps domains
Interoperability with Google Apps For Your Domain can cause headaches;
if the administrator fails to add the requisite SRV records,
federation with other domains will obviously fail.  However, they will
be able to chat with other Google Talk users, so jabber will appear to
"work" from their point of view.

This adds a configuration option to delegate a SRV lookup to
gmail.com, when a domain lacking SRV records is likely a Google Apps
domain (judged based on the MX records on it).  As this option is a
heuristic, and attempts to compensate for flaws in other domains, it
is disabled by default.